Lots of new functions today. The primary new functions are any within the crown function family and the stand density index family.
the previous merc function was renamed and
should be fixed in everything where it is a dependency. smalians was changed to Smalians.

Function documentation improved for almost the entire package. Examples were added, seealso
was included, family was included where appropriate.

None of the functions are properly cited.

There are likely errors. The CCF and SDI.Plot functions are giving very high values.